+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * 2.0.0 February 15, 2017
  - Add eslint-plugin-mocha and mocha/no-mocha-arrows rule
+ - Disable prefer-arrow-callback rule
 
 * 1.0.3 January 18, 2017
   - Fixup dependencies